// public/pages/ForecastDetail/utils/interface.ts
/*
 * SPDX-License-Identifier: Apache-2.0
 *
 * The OpenSearch Contributors require contributions made to
 * this file be licensed under the Apache-2.0 license or a
 * compatible open source license.
 *
 * Modifications Copyright OpenSearch Contributors. See
 * GitHub history for details.
 */

export interface DateRange {
    startDate: string | number; // Can be relative ('now-1h') or absolute (epoch ms)
    endDate: string | number;   // Can be relative ('now') or absolute (epoch ms) 
    isRelative?: boolean;       // Flag to indicate if using relative dates
}